A large set of edits aimed at making the system more compatible with full 64-bit mode, such as avoiding truncation of 64-bit  offsets, etc.  Fixed a bunch of Makefiles which did not pass down the arch flags to their link commands, which caused many executables to fail to link.

+++ insts/std/LPCPLAY/DataSet.h
@@ -6,20 +6,20 @@ class DataSet : public RefCounted
 {
 public:
 	DataSet();
-	int	open(const char *fileName, int npoleGuess, float sampRate);
+	off_t	open(const char *fileName, int npoleGuess, float sampRate);
 	inline int getNPoles() { return _nPoles; }
-	inline int getFrameCount() { return _frameCount; }
+	inline off_t getFrameCount() { return _frameCount; }
 	int	getFrame(float frameno, float *pCoeffs);
 protected:
 	~DataSet();
 	void	allocArray(int nPoles);
 private:
 	int	_nPoles;
-	int _frameCount;
+	off_t _frameCount;
 	int	_fdesc;
 	int	_lpHeaderSize;
 	float	*_array;
-	int	_oldframe, _endframe;
+	off_t	_oldframe, _endframe;
 	int	_framsize;
 	int	_fprec;
 	int	_recsize;
